reverting rerender fix in BasicKeyframedTrack
