+++ theatre/studio/src/panels/SequenceEditorPanel/DopeSheet/Right/BasicKeyframedTrack/BasicKeyframedTrack.tsx
@@ -1,44 +1,35 @@
-import React, {useState} from 'react'
 import type {TrackData} from '@theatre/core/projects/store/types/SheetState_Historic'
-import type {
-  DopeSheetSelection,
-  SequenceEditorPanelLayout,
-} from '@theatre/studio/panels/SequenceEditorPanel/layout/layout'
+import type {SequenceEditorPanelLayout} from '@theatre/studio/panels/SequenceEditorPanel/layout/layout'
 import type {SequenceEditorTree_PrimitiveProp} from '@theatre/studio/panels/SequenceEditorPanel/layout/tree'
 import {usePrism} from '@theatre/react'
 import type {Pointer} from '@theatre/dataverse'
 import {val} from '@theatre/dataverse'
+import React from 'react'
 import KeyframeEditor from './KeyframeEditor/KeyframeEditor'
-import type {StrictRecord} from '@theatre/shared/utils/types'
-import isEqual from 'lodash-es/isEqual'
 
 const BasicKeyframedTrack: React.FC<{
   leaf: SequenceEditorTree_PrimitiveProp
+
   layoutP: Pointer<SequenceEditorPanelLayout>
   trackData: TrackData
 }> = React.memo(({layoutP, trackData, leaf}) => {
-  const [selection, setSelection] = useState<undefined | DopeSheetSelection>(
-    undefined,
-  )
-  const [selectedKeyframeIds, setSelectedKeyframeIds] = useState<
-    undefined | StrictRecord<string, true>
-  >({})
-
-  usePrism(() => {
+  // TODO: Prevent all the rerenders this causes when selecting keyframes
+  const {selectedKeyframeIds, selection} = usePrism(() => {
     const selectionAtom = val(layoutP.selectionAtom)
-    const newSelectedKeyframeIds =
-      val(
-        selectionAtom.pointer.current.byObjectKey[
-          leaf.sheetObject.address.objectKey
-        ].byTrackId[leaf.trackId].byKeyframeId,
-      ) || {}
-
-    // Only update these objects when selectedKeyframeIds change
-    if (!isEqual(selectedKeyframeIds, newSelectedKeyframeIds)) {
-      setSelectedKeyframeIds(newSelectedKeyframeIds)
-      setSelection(val(selectionAtom.pointer.current) || undefined)
+    const selectedKeyframeIds = val(
+      selectionAtom.pointer.current.byObjectKey[
+        leaf.sheetObject.address.objectKey
+      ].byTrackId[leaf.trackId].byKeyframeId,
+    )
+    if (selectedKeyframeIds) {
+      return {
+        selectedKeyframeIds,
+        selection: val(selectionAtom.pointer.current),
+      }
+    } else {
+      return {selectedKeyframeIds: {}, selection: undefined}
     }
-  }, [layoutP, leaf.trackId, selection, selectedKeyframeIds])
+  }, [layoutP, leaf.trackId])
 
   const keyframeEditors = trackData.keyframes.map((kf, index) => (
     <KeyframeEditor
